cli update
